+++ apps/prairielearn/src/lib/markdown.js
@@ -1,3 +1,4 @@
+//@ts-check
 const unified = require('unified');
 const markdown = require('remark-parse');
 const raw = require('rehype-raw');
@@ -17,6 +18,7 @@ const langRegex = /([^\\{]*)?(\{(.*)\})?/;
 
 const visitCodeBlock = (ast, _vFile) => {
   return visit(ast, 'code', (node, index, parent) => {
+    // @ts-expect-error - TODO: resolve after converting file to TypeScript. Currently, node.lang & node.value do not exist on type Node<Data>.
     let { lang, value } = node;
     const attrs = [];
 
@@ -38,10 +40,9 @@ const visitCodeBlock = (ast, _vFile) => {
       type: 'html',
       value: `<pl-code ${attrs.join(' ')}>${value}</pl-code>`,
     };
+    parent?.children.splice(index, 1, html);
 
-    parent.children.splice(index, 1, html);
-
-    return node;
+    return;
   });
 };
 
@@ -53,10 +54,12 @@ const visitCodeBlock = (ast, _vFile) => {
  */
 const visitCheckSingleParagraph = (ast, _vFile) => {
   return visit(ast, 'root', (node, _index, _parent) => {
+    // @ts-expect-error - TODO: resolve after converting file to TypeScript
     if (node.children.length === 1 && node.children[0].tagName === 'p') {
+      // @ts-expect-error - TODO: resolve after converting file to TypeScript. Currently, node.children does not exist on type Node<Data>.
       node.children = node.children[0].children;
     }
-    return node;
+    return;
   });
 };
 
@@ -73,10 +76,11 @@ const visitMathBlock = (ast, _vFile) => {
     const endFence = node.type === 'math' ? '\n$$' : '$';
     const text = {
       type: 'text',
+      // @ts-expect-error - TODO: resolve after converting file to TypeScript. Currently, node.value does not exist on type Node<Data>.
       value: startFence + node.value + endFence,
     };
-    parent.children.splice(index, 1, text);
-    return node;
+    parent?.children.splice(index, 1, text);
+    return;
   });
 };
 
@@ -127,7 +131,7 @@ const questionProcessor = unified()
   .use(raw)
   .use(stringify);
 
-module.exports.processQuestion = function (html) {
+export function processQuestion(html) {
   return html.replace(regex, (_match, originalContents) => {
     // We'll handle escapes before we pass off the string to our Markdown pipeline
     const decodedContents = originalContents.replace(escapeRegex, (match, prefix, hashes) => {
@@ -136,16 +140,16 @@ module.exports.processQuestion = function (html) {
     const res = questionProcessor.processSync(decodedContents);
     return res.contents;
   });
-};
+}
 
-module.exports.processContent = async function (original) {
+export async function processContent(original) {
   return (await defaultProcessor.process(original)).contents;
-};
+}
 
 /**
  * This function is similar to `processContent`, except that if the content fits a single line
- * (paragrah) it will return the content without a `p` tag.
+ * (paragraph) it will return the content without a `p` tag.
  */
-module.exports.processContentInline = async function (original) {
+export async function processContentInline(original) {
   return (await inlineProcessor.process(original)).contents;
-};
+}