+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/sentry
 
+## 1.1.5
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+
+## 1.1.4
+
+### Patch Changes
+
+- 2da23ab: Upgrade all JavaScript dependencies
+
+## 1.1.3
+
+### Patch Changes
+
+- 1523b97b0: Upgrade all dependencies
+
 ## 1.1.2
 
 ### Patch Changes